requirements.txt
python-binance>=1.0.19
python-dotenv>=1.0.0
streamlit>=1.28.0
typer>=0.9.0
rich>=13.6.0

